Input: synaptics-rmi4: i2c: factor out initial power up
To support being a panel follower, we want to be able to optionally
defer powering up and first communicating with the touchscreen until
after the associated panel is powered up. In preparation for that,
factor the actual power up out of the probe function into a helper, and
call it last of all in the probe function.