Input: synaptics-rmi4: i2c: move suspend and resume to helper functions
In a future change to add support for being a DRM panel follower, we'd
like to call suspend and resume functions when the panel is unpreparing
and prepared respectively. Factor out the suspend and resume mechanics
into helper functions taking a struct rmi_i2c_xport to make this easier.

The existing system sleep PM functions and runtime PM functions can both
use the new helper functions as well.